do not update display if tag stack has not changed
It would be more efficient to check if tagstack.curidx has changed, but
that is incorrect since the SafeState event will not fire between
consecutive commands. For example,

:pop | cs find g <symbol_name>

will change the tag stack, but won't update the display if only changes
to curidx are checked.

Despite not being the most efficient solution, some crude testing
indicates this change does reduce CPU usage of the terminal emulator by
about 25%.